Moved static class used to fetch parameters
- Moved class used to fetch GET parameters.
